refactor: postApiAction으로 중복되는 코드 제거
