+++ src/pages/Teachers.tsx
@@ -5,11 +5,10 @@ import Wrapper from "../components/common/Wrapper";
 import { useEffect } from "react";
 import useCategory from "../hooks/useCategory";
 import usePostScrollPage from "../hooks/usePostScrollPage";
-import { getPost } from "../store/post/api/PostReadThunk";
+import { getPost } from "../store/post/postApiAction";
 import { useDispatch } from "react-redux";
 import { ThunkDispatch } from "@reduxjs/toolkit";
 import { setCategory } from "../store/category";
-import { useParams } from "react-router-dom";
 
 const Container = styled.div`
   margin-top: 4rem;
@@ -22,20 +21,20 @@ const Container = styled.div`
 
 export default function Teachers() {
   const [page, setPage] = usePostScrollPage();
-  const [category, categoryDispatch] = useCategory();
-  const p = useParams();
-  console.log(p);
+  const [, categoryDispatch] = useCategory();
 
   const dispatch = useDispatch<ThunkDispatch<any, void, any>>();
 
   useEffect(() => {
+    // console.log("페이지 처음 마운트", page);
     setPage(0);
     categoryDispatch(setCategory("teacher"));
   }, [categoryDispatch, setPage]);
 
   useEffect(() => {
-    dispatch(getPost({ category, page }));
-  }, [dispatch, category, page]);
+    // console.log("이제야 제대로", page);
+    dispatch(getPost({ category: "teacher", page }));
+  }, [dispatch, page]);
 
   return (
     <Wrapper>
